requirements.txt
django
djangorestframework
django-cors-headers
django-oauth-toolkit
pymemcache
Pillow
python-dotenv
django-filter
pyotp
django-cacheops